Scaling implementation
Added the basic scaling to the canvas, based on the scroll movement. What's more, I've started to work on a system that creates non-square canvas shapes; it's more work than I was thinking originally, but it's worth a time to implement this. The main problem is probably in the 'Layer' class in the 'SetPixelColor' function with it's 2D coordinates to 1D vector calculation.