Core program systems.
Created the first mandatory systems for Colorful Pixels: Layer, Color and Tool systems. Canvas was, in some way at least, modified to be it's own system as well. What's more, in 'ColorfulPixels.hpp' header there are macros with the current version of the program (right now: Alpha 1.0). Addition of the new tool: Color Picker, which gets the color of the hovered cell and switches to the brush tool. What's more, now you can select which tool do you want to use via the keyboard shortcuts. Slowly getting to the point where I can publish the software as an alpha demo, which then I want to develop further to be it's own small 2D pixel art editor.