Tinyfiledialogs implementation
Implemented tinyfiledialogs into the project. Now saving and loading the .cfpx file (json with different extension) is much more easier and comfortable. What's more, debugging was tweaked to suite the style of raylib's debug logging ('MODULE: Message'). Garbage files were removed and a small  was added; for now it's just unseting the PATH variables. Added some overloads to LayerSystem. Cursor is now colored the same as the currently used color. Now all that's left is to clean-up the codebase, find and get rid of bugs, fix the problem with WEB version, creating an icon and releasing the software to the public.